simplify regression modules
