+++ torchmetrics/classification/hamming_distance.py
@@ -16,7 +16,7 @@ from typing import Any, Callable, Optional
 import torch
 from torch import Tensor, tensor
 
-from torchmetrics.functional.classification.hamming_distance import _hamming_distance_compute, _hamming_distance_update
+from torchmetrics.functional.classification.hamming import _hamming_distance_compute, _hamming_distance_update
 from torchmetrics.metric import Metric
