+++ torchmetrics/functional/classification/__init__.py
@@ -20,7 +20,7 @@ from torchmetrics.functional.classification.cohen_kappa import cohen_kappa  # no
 from torchmetrics.functional.classification.confusion_matrix import confusion_matrix  # noqa: F401
 from torchmetrics.functional.classification.dice import dice_score  # noqa: F401
 from torchmetrics.functional.classification.f_beta import f1, f1_score, fbeta, fbeta_score  # noqa: F401
-from torchmetrics.functional.classification.hamming_distance import hamming_distance  # noqa: F401
+from torchmetrics.functional.classification.hamming import hamming_distance  # noqa: F401
 from torchmetrics.functional.classification.hinge import hinge, hinge_loss  # noqa: F401
 from torchmetrics.functional.classification.jaccard import jaccard_index  # noqa: F401
 from torchmetrics.functional.classification.kl_divergence import kl_divergence  # noqa: F401
